fp8 kv cache is not working

+++ python/minisgl/attention/fa.py
@@ -81,7 +81,7 @@ class FlashAttentionBackend(BaseAttnBackend):
         # print(k_cache[:metadata.cache_seqlens[0]])
 
         return flash3_kvcache_impl(
-            q_flash,
+            q_flash.to(k_cache.dtype),
             k_cache,
             v_cache,
             cache_seqlens=metadata.cache_seqlens,
